feat(vmm): add vhost-user-blk device
This adds vhost-user block device implementation.

In order to create a vhost-user block device,
a PUT /drives request needs to be issued that
contains `vhost-user` object.
At the moment, the object has the only property `socket`
that is a socket path to communicate with
a vhost-user backend.

Example:
```
curl --unix-socket ${fc_socket} -i \
  -X PUT "http://localhost/drives/vhost" \
  -H "accept: application/json" \
  -H "Content-Type: application/json" \
  -d "{
            \"drive_id\": \"vhost\",
            \"is_root_device\": false,
            \"cache_type\": \"Unsafe\",
            \"vhost_user\": {
              \"socket\": \"${vhost_socket}\"
            }
      }"
```

+++ src/vmm/src/devices/virtio/vhost_user/mod.rs
@@ -182,7 +182,7 @@ impl VhostUserHandle {
 
             let vhost_user_net_reg = VhostUserMemoryRegionInfo {
                 guest_phys_addr: region.start_addr().raw_value(),
-                memory_size: region.len() as u64,
+                memory_size: region.len(),
                 userspace_addr: region.as_ptr() as u64,
                 mmap_offset,
                 mmap_handle,
@@ -209,6 +209,9 @@ impl VhostUserHandle {
         self.vu
             .set_features(acked_features)
             .map_err(VhostUserError::VhostUserSetFeatures)?;
+
+        // Update acked features after they have been sent to the backend.
+        // self.acked_features = acked_features;
 
         // Provide the memory table to the backend.
         self.update_mem_table(mem)?;
